add download poit as an git-repo, pull for update

+++ openpoti/cli.py
@@ -1,25 +1,28 @@
-'''
-TODO:
-1. Download catalog.csv
-2. Download all the OpenPotis
-3. poti layer -n <layer-name> -l <layer_01, layer_02, ...> -a -o <path to opf file>
-'''
 import csv
 import requests
 import zipfile
 from io import BytesIO, StringIO
 from pathlib import Path
+import shutil
 
 import click
+from github import Github
+from git import Repo
 from tqdm import tqdm
 
 from openpoti.serializemd import SerializeMd
 
-
+OP_PATH = './.openpoti'
 config = {
-    'CATALOG_URL': 'https://raw.githubusercontent.com/OpenPoti/openpoti-catalog/master/data/catalog.csv',
-    'OpenPotiRepo': 'https://github.com/OpenPoti/',
-    'CONFIG_PATH': './.openpoti/config/data_config'
+    # Github
+    'OP_CATALOG_URL': 'https://raw.githubusercontent.com/OpenPoti/openpoti-catalog/master/data/catalog.csv',
+    'OP_ORG': 'https://github.com/OpenPoti',
+
+    # Local
+    'OP_DATA_PATH': f'{OP_PATH}/data',
+    'OP_CATALOG_PATH': f'{OP_PATH}/data/catalog.csv',
+    'CONFIG_PATH': f'{OP_PATH}/config',
+    'DATA_CONFIG_PATH': f'{OP_PATH}/config/data_config',
 }
 
 
@@ -28,56 +31,71 @@ def cli():
     pass
 
 
+def create_config_dirs():
+    Path(config['OP_DATA_PATH']).mkdir(parents=True, exist_ok=True)
+    Path(config['CONFIG_PATH']).mkdir(parents=True, exist_ok=True)
+
+
 def get_poti_list():
     poti_list = []
 
-    # download the catalog
-    r = requests.get(config['CATALOG_URL'])
-    f = StringIO(r.content.decode('utf-8'))
+    catalog_path = Path(config['OP_CATALOG_PATH'])
+    if not catalog_path.is_file():
+        # download the catalog
+        r = requests.get(config['OP_CATALOG_URL'])
+        catalog_path.write_text(r.content.decode('utf-8'))
     
     # parse the poti workid
-    reader = csv.reader(f, delimiter=',')
-    for row in reader:
-        if not row[0].startswith('W'): continue
-        poti_list.append(row[0])
+    with catalog_path.open('r') as f:
+        reader = csv.reader(f, delimiter=',')
+        for row in reader:
+            if not row[0].startswith('W'): continue
+            poti_list.append(row[0])
 
     return poti_list
 
 
-def download_poti(poti):
-    poti_url = f'https://github.com/OpenPoti/{poti}/archive/master.zip'
-
-    # download the poti
-    r = requests.get(poti_url)
-    z = zipfile.ZipFile(BytesIO(r.content))
-    z.extractall(config['data'])
-
-    # rename the poti
-    target = config['data']/poti
-    if not target.is_dir():
-        source = config['data']/f'{poti}-master'
-        source.replace(target)
-
-    #TODO: updating the poti
+def download_poti(poti, out):
+    # clone the repo
+    poti_url = f"{config['OP_ORG']}/{poti}.git"
+    poti_path = Path(config['OP_DATA_PATH'])/poti
+    if poti_path.is_dir(): # if repo is already exits at local then try to pull
+        repo = Repo(str(poti_path))
+        repo.remotes.origin.pull()
+    else:
+        Repo.clone_from(poti_url, str(poti_path))
+    
+    # Create duplicate base-text for editing
+    base_text = poti_path/f'{poti}.opf'/'base.txt'
+    out_dir = Path(out)
+    if not out_dir.is_dir(): out_dir.mkdir()
+    dup_base_text = Path(out)/f'{poti}.txt'
+    shutil.copy(str(base_text), str(dup_base_text))
 
 
 # Download openPoti
 @cli.command()
-@click.option('--save_dir', default='./.openpoti/data',
+@click.option('--out', '-o', default='./poti',
                             help='directory to store all the poti')
-def download(save_dir):
+def download(**kwargs):
+    # create config dirs
+    create_config_dirs()
+
     # configure the data_path
-    config['data'] = Path(save_dir)
+    config['data'] = Path(kwargs['out'])
 
     # download the repo
     poti_list = get_poti_list()
-    for poti in tqdm(poti_list):
-        download_poti(poti)
+    for poti in tqdm(poti_list[:2]):
+        download_poti(poti, kwargs['out'])
 
     # save data_path in data_config
-    config_path = Path(config['CONFIG_PATH'])
-    config_path.parent.mkdir(parents=True, exist_ok=True)
-    config_path.write_text(str(config['data'].resolve()))
+    config_path = Path(config['DATA_CONFIG_PATH'])
+    if not config_path.is_file():
+        config_path.write_text(str(config['data'].resolve()))
+
+    # print location of data
+    print('Poti saved at:', Path(kwargs['out']).resolve())
 
 
 # Apply layer
@@ -92,7 +110,7 @@ layers_name = ['title', 'tsawa', 'yigchung', 'quotes', 'sapche']
 @click.argument('out', type=click.File('w'))
 def layer(**kwargs):
     """
-    Command to apply a single layer, multiple layers or all available layers, which is the default.\n
+    Command to apply a single layer, multiple layers or all available layers (by default) and then export to markdown.\n
     Args:\n
         - OPF_PATH is the path to opf directory of poti\n
         - OUT is the filename to the write the result. Currently support only Markdown file.
